Add support for non-string extend/override proto values
Involves changing the way we're building the method call expressions, since we no longer just have a string representation of the class to contend with - it may be an actual identifier/member expression.

This also meant it was easier to use the name of the class for the unique variable name when doing a method override, since we can be sure that will be a string.