Massive refactor to move all 'Definition' specific code to a new file
This moves all of the code for finding the `Ext.define` or `Ext.override` definition into a new file, and has some big refactoring of how this code is structured.

There is now a clear separation between the code paths for when we're in `Ext.define` and when we're in `Ext.override`, and each will return a `Definition` instance of either `DefineDefinition` or `OverrideDefinition` respectively.

This means that the codepath for `OverrideDefinition` can skip out a lot of the inspection for the prototype property, and determining whether it's an `extend` or `override` prototype (it knows it's always an override).

Both `Definition` classes extend a common abstract which handles the common requirement for prepending the method reference before the definition when it is an override.